MutableArithmetics for IPM/HSD
Use the `BigFloat` dot product from MutableArithmetics in HSD code.

Helps with the performance of the `BigFloat` arithmetic. The change
shouldn't affect other arithmetics, and it's coded so it'd be easy to
extend it to another mutable arithmetic apart from just `BigFloat`, if
necessary, and if such a type will support MutableArithmetics.

Apart from improving performance, this change could possibly also
benefit LP problems with numerical issues (when using `BigFloat`),
because the MA dot product uses a summation algorithm that's more
accurate than naive summation.

A performance experiment is presented in the commit message of the
following commit. The conclusion is that this commit improves
performance only by a tiny bit, likewise with allocation.